requirements.txt
backports.entry-points-selectable==1.1.0
click==8.0.3
colorama==0.4.4
distlib==0.3.3
filelock==3.3.0
Flask==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
platformdirs==2.4.0
six==1.16.0
typing-extensions==3.10.0.2
virtualenv==20.8.1
Werkzeug==2.0.2
zipp==3.6.0
